+++ CHANGES.rst
@@ -15,6 +15,7 @@ Changelog
 - Extended the _notifyContainerModified event to default Plone elements Folder/File/Document/News Item, so when using a 'Documents' folder to publish some documents,
   adding a new element will notify container modified and invalidate cache
 - Added adaptable method MeetingItem.custom_validate_optionalAdvisers so a plugin may validate selected optional advisers if necessary
+- Display asked advices on the meetingitem_view at top left together with copy groups so informations about who may see the item is located at the same place
 
 4.1rc9 (2019-08-13)
 -------------------